% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clusters.R
\name{create_clusters}
\alias{create_clusters}
\title{Creates clustering of data}
\usage{
create_clusters(cols, num)
}
\arguments{
\item{cols}{Columns (fields) of a data frame}

\item{num}{Number of clusters}
}
\value{
kmeans
}
\description{
Takes columns and number of clusters. Runs the basic R kmeans
function and returns the output.
}
