% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plots two-dimensional clusters}
\usage{
plot_clusters(data, xcol, ycol, kmeans)
}
\arguments{
\item{kmeans}{Result of the kmeans algorithm}

\item{cols}{The columns to be plotted}
}
\value{
plot
}
\description{
Takes two column vectors with their associated clusters and plots them
on a standard dot plot. Also plots the centers of each cluster.
}
